[MOV] split_to_columns: move store file to correct folder
Task: 6394556
